+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 #develop
   * CHANGE - Menu intros are one message, making them skippable.
+  * FEATURE - Personalized voicemail greetings can be deleted
 
 #v1.0.0 - 2014-01-02
   * FEATURE - Optional per-mailbox override of default recording hash